update build tag for example/
